+++ Cargo.toml
@@ -32,8 +32,8 @@ members = [
 
     "consensus/cached_tree_hash",
     "consensus/int_to_bytes",
-    "consensus/lmd_ghost",
-    "consensus/proto_array_fork_choice",
+    "consensus/fork_choice",
+    "consensus/proto_array",
     "consensus/safe_arith",
     "consensus/ssz",
     "consensus/ssz_derive",
