+++ common/remote_beacon_node/Cargo.toml
@@ -17,5 +17,5 @@ hex = "0.4.2"
 eth2_ssz = "0.1.2"
 serde_json = "1.0.52"
 eth2_config = { path = "../eth2_config" }
-proto_array_fork_choice = { path = "../../consensus/proto_array_fork_choice" }
+proto_array = { path = "../../consensus/proto_array" }
 operation_pool = { path = "../../beacon_node/operation_pool" }
