+++ consensus/fork_choice/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "lmd_ghost"
+name = "fork_choice"
 version = "0.1.0"
 authors = ["Paul Hauner <paul@paulhauner.com>"]
 edition = "2018"
@@ -8,7 +8,7 @@ edition = "2018"
 
 [dependencies]
 types = { path = "../types" }
-proto_array_fork_choice = { path = "../proto_array_fork_choice" }
+proto_array = { path = "../proto_array" }
 eth2_ssz = { path = "../ssz" }
 eth2_ssz_derive = { path = "../ssz_derive" }
