+++ consensus/proto_array/Cargo.toml
@@ -1,11 +1,11 @@
 [package]
-name = "proto_array_fork_choice"
+name = "proto_array"
 version = "0.2.0"
 authors = ["Paul Hauner <paul@sigmaprime.io>"]
 edition = "2018"
 
 [[bin]]
-name = "proto_array_fork_choice"
+name = "proto_array"
 path = "src/bin.rs"
 
 [dependencies]
