+++ consensus/proto_array/src/bin.rs
@@ -1,4 +1,4 @@
-use proto_array_fork_choice::fork_choice_test_definition::*;
+use proto_array::fork_choice_test_definition::*;
 use serde_yaml;
 use std::fs::File;
