+++ consensus/proto_array/src/error.rs
